Move commands to pkg/commands and split into files.
