+++ pkg/commands/publisher_test.go
